refactor: 입력부 문자열split 로직 책임분리
